requirements.txt
dotenv==0.9.9
numpy==2.4.2
pandas==3.0.0
psycopg2-binary==2.9.11
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
